Created a stabilized video stitcher
Split the global frame drawer into two sub functions, the first reads the file, determining the transformation between frames and calculating the required size for the global frame, while the second reads the file again, taking the pre-deteremined transformation and calculated parameters as arguments to stitch together a video with a fixed global coordinate system.

The result works but shows room for improvement.

Also adjusted the detection to track assignment function to take the cost of non assignment as a parameter so that it may be changed under different conditions such as scene changes.